fixes no existing file

+++ retrieve_data/service_isFace.py
@@ -11,7 +11,8 @@ def exec_isFace_func():
 	if request.method == 'POST':
 		f = request.files['']
 		os.makedirs(save_folder_path, exist_ok=True)
-		os.remove(save_file_path)
+		if os.path.exists(save_file_path):
+			os.remove(save_file_path)
 		f.save(save_file_path)
 		(n_faces, scores) = calc_conf_score(save_file_path, isPath=True)
 		return jsonify(number_of_faces=n_faces,
